requirements.txt
python-telegram-bot
pandas
matplotlib
openpyxl
apscheduler